learn_fastapi/pyproject.toml
[tool.poetry]
name = "learn-fastapi"
version = "0.1.0"
description = ""
authors = ["dima cheban <dimacheban23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.6"}
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.37"
pydantic-settings = "^2.7.1"
poetry-dotenv-plugin = "^0.2.0"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
alembic = "^1.14.0"
pytest = "^8.3.4"
aiohttp = "^3.11.11"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
